[minor] fix all the signed/unsigned comparison warnings
I tried to apply the two following principles relatively uniformly:
- if a variable is obviously an array index and nothing more, use size_t
- if a cast in needed anyway, because the loop variable is really used
  as an int, favor signed ints and static_cast<int> the unsigned ones

+++ Goblin Camp/include/tileRenderer/Sprite.hpp
@@ -94,8 +94,8 @@ template <typename IterT> Sprite::Sprite(IterT start, IterT end, bool connection
 		return;
 	}
 
-	if (frameCount > indices.size()) {
-		frameCount = indices.size();
+	if (frameCount > static_cast<int>(indices.size())) {
+		frameCount = static_cast<int>(indices.size());
 	}
 	int numTiles = indices.size() / frameCount;
 	if (numTiles == 0) {
